created background/foreground class
Also began implementation for said class. More changes to come soon.
