--- Source/foreBackgroundHandler.cpp
@@ -1,4 +1,0 @@
-#include "foreBackgroundHandler.h"
-
-vector<entity*> background;
-vector<entity*> foreground;
