+++ Source/ground.cpp
@@ -1,0 +1,4 @@
+#include "ground.h"
+
+vector<ground*> background;
+vector<ground*> foreground;
